Conform to og:title requirements
According to http://ogp.me/, og:title should only include the name of
the item, and not the site name.

So, create a `og_title` block that allows setting just the name of the
item, and include that in the base template if it exists. Fall back to
`title` if `og_title` was not specified.